Refactor internals: move time calculations logic to a different module
