+++ lib/jekyll-timeago.rb
@@ -1,3 +1,4 @@
+require 'jekyll-timeago/core'
 require 'jekyll-timeago/filter'
 require 'jekyll-timeago/tag'
 require 'jekyll-timeago/version'
