+++ script/console
@@ -3,8 +3,8 @@
 lib = File.expand_path('../../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
 
-require 'jekyll-timeago/filter'
-include Jekyll::Timeago::Filter
+require 'jekyll-timeago/core'
+include Jekyll::Timeago::Core
 
 require 'active_support/all'
